RavenDB-18513 - Addressed PR Comments
