+++ test/SlowTests/Issues/RavenDB-18513.cs
@@ -32,7 +32,7 @@ public class RavenDB_18513 : RavenTestBase
     }
 
     [RavenFact(RavenTestCategory.Revisions)]
-    public async Task Test()
+    public async Task RevertRevisionShouldntCreateOrphanedRevisions()
     {
         DateTime beforeStore = DateTime.UtcNow - TimeSpan.FromDays(1);
 
@@ -67,7 +67,7 @@ public class RavenDB_18513 : RavenTestBase
         var empyConfiguration = new RevisionsConfiguration();
         await RevisionsHelper.SetupRevisions(store, Server.ServerStore, configuration: empyConfiguration);
         // revert revision
-        var operation = await store.Maintenance.SendAsync(new RevertRevisionsOperation(beforeStore, 60));
+        var operation = await store.Maintenance.SendAsync(new RevisionsHelper.RevertRevisionsOperation(beforeStore, 60));
         await operation.WaitForCompletionAsync(TimeSpan.FromSeconds(15)).ConfigureAwait(false);
 
         using (var session = store.OpenAsyncSession())
@@ -94,57 +94,6 @@ public class RavenDB_18513 : RavenTestBase
     {
         public string Id { get; set; }
         public string Name { get; set; }
-    }
-
-    private class RevertRevisionsOperation : IMaintenanceOperation<OperationIdResult>
-    {
-        private readonly RevertRevisionsRequest _request;
-
-        public RevertRevisionsOperation(DateTime time, long window)
-        {
-            _request = new RevertRevisionsRequest() { Time = time, WindowInSec = window };
-        }
-
-        public RevertRevisionsOperation(RevertRevisionsRequest request)
-        {
-            _request = request ?? throw new ArgumentNullException(nameof(request));
-        }
-
-        public RavenCommand<OperationIdResult> GetCommand(DocumentConventions conventions, JsonOperationContext context)
-        {
-            return new RevertRevisionsCommand(_request);
-        }
-
-        private class RevertRevisionsCommand : RavenCommand<OperationIdResult>
-        {
-            private readonly RevertRevisionsRequest _request;
-
-            public RevertRevisionsCommand(RevertRevisionsRequest request)
-            {
-                _request = request;
-            }
-
-            public override bool IsReadRequest => false;
-
-            public override HttpRequestMessage CreateRequest(JsonOperationContext ctx, ServerNode node, out string url)
-            {
-                url = $"{node.Url}/databases/{node.Database}/revisions/revert";
-
-                return new HttpRequestMessage
-                {
-                    Method = HttpMethod.Post,
-                    Content = new BlittableJsonContent(async stream => await ctx.WriteAsync(stream, DocumentConventions.Default.Serialization.DefaultConverter.ToBlittable(_request, ctx)).ConfigureAwait(false))
-                };
-            }
-
-            public override void SetResponse(JsonOperationContext context, BlittableJsonReaderObject response, bool fromCache)
-            {
-                if (response == null)
-                    ThrowInvalidResponse();
-
-                Result = DocumentConventions.Default.Serialization.DefaultConverter.FromBlittable<OperationIdResult>(response);
-            }
-        }
     }
 }
