+++ test/SlowTests/Server/Documents/Revisions/RevertRevisionsByCollectionTests.cs
@@ -463,7 +463,7 @@ namespace SlowTests.Server.Documents.Revisions
                     await session.SaveChangesAsync();
                 }
 
-                var operation = await store.Maintenance.SendAsync(new RevertRevisionsOperation(last, 60, collections));
+                var operation = await store.Maintenance.SendAsync(new RevisionsHelper.RevertRevisionsOperation(last, 60, collections));
                 await operation.WaitForCompletionAsync(TimeSpan.FromSeconds(5)).ConfigureAwait(false);
 
                 using (var session = store.OpenAsyncSession())
@@ -525,7 +525,7 @@ namespace SlowTests.Server.Documents.Revisions
                     await session.SaveChangesAsync();
                 }
 
-                var operation = await store.Maintenance.SendAsync(new RevertRevisionsOperation(last, 60, collections));
+                var operation = await store.Maintenance.SendAsync(new RevisionsHelper.RevertRevisionsOperation(last, 60, collections));
                 await operation.WaitForCompletionAsync(TimeSpan.FromSeconds(5)).ConfigureAwait(false);
 
                 using (var session = store.OpenAsyncSession())
@@ -576,7 +576,7 @@ namespace SlowTests.Server.Documents.Revisions
                     await session.SaveChangesAsync();
                 }
 
-                var operation = await store.Maintenance.SendAsync(new RevertRevisionsOperation(last, 60, collections));
+                var operation = await store.Maintenance.SendAsync(new RevisionsHelper.RevertRevisionsOperation(last, 60, collections));
                 await operation.WaitForCompletionAsync(TimeSpan.FromSeconds(5)).ConfigureAwait(false);
 
                 using (var session = store.OpenAsyncSession())
@@ -642,7 +642,7 @@ namespace SlowTests.Server.Documents.Revisions
                     session.SaveChanges();
                 }
 
-                var operation = await store.Maintenance.SendAsync(new RevertRevisionsOperation(last, 60, collections));
+                var operation = await store.Maintenance.SendAsync(new RevisionsHelper.RevertRevisionsOperation(last, 60, collections));
                 await operation.WaitForCompletionAsync(TimeSpan.FromSeconds(5)).ConfigureAwait(false);
 
                 using (var session = store.OpenAsyncSession())
@@ -700,7 +700,7 @@ namespace SlowTests.Server.Documents.Revisions
 
                 var db = await Databases.GetDocumentDatabaseInstanceFor(store);
 
-                var operation = await store.Maintenance.SendAsync(new RevertRevisionsOperation(last, 60, collections)); 
+                var operation = await store.Maintenance.SendAsync(new RevisionsHelper.RevertRevisionsOperation(last, 60, collections)); 
                 var result = await operation.WaitForCompletionAsync(TimeSpan.FromSeconds(5)).ConfigureAwait(false);
 
                 using (var session = store.OpenAsyncSession())
@@ -741,7 +741,7 @@ namespace SlowTests.Server.Documents.Revisions
                     await session.SaveChangesAsync();
                 }
 
-                var operation = await store.Maintenance.SendAsync(new RevertRevisionsOperation(last, 60));
+                var operation = await store.Maintenance.SendAsync(new RevisionsHelper.RevertRevisionsOperation(last, 60));
                 await operation.WaitForCompletionAsync(TimeSpan.FromSeconds(5)).ConfigureAwait(false);
 
                 using (var session = store.OpenAsyncSession())
@@ -759,67 +759,6 @@ namespace SlowTests.Server.Documents.Revisions
                     Assert.Equal("User Name", usersRevisions[0].Name);
                     Assert.Equal("Shahar", usersRevisions[1].Name);
                     Assert.Equal("User Name", usersRevisions[2].Name);
-                }
-            }
-        }
-
-        private class RevertRevisionsOperation : IMaintenanceOperation<OperationIdResult>
-        {
-            private readonly RevertRevisionsRequest _request;
-
-            public RevertRevisionsOperation(DateTime time, long window)
-            {
-                _request = new RevertRevisionsRequest() { 
-                    Time = time, 
-                    WindowInSec = window,
-                };
-            }
-
-            public RevertRevisionsOperation(DateTime time, long window, string[] collections) : this(time, window)
-            {
-                if(collections == null || collections.Length == 0)
-                    throw new InvalidOperationException("Collections array cant be null or empty.");
-                _request.Collections = collections;
-            }
-
-            public RevertRevisionsOperation(RevertRevisionsRequest request)
-            {
-                _request = request ?? throw new ArgumentNullException(nameof(request));
-            }
-
-            public RavenCommand<OperationIdResult> GetCommand(DocumentConventions conventions, JsonOperationContext context)
-            {
-                return new RevertRevisionsCommand(_request);
-            }
-
-            private class RevertRevisionsCommand : RavenCommand<OperationIdResult>
-            {
-                private readonly RevertRevisionsRequest _request;
-
-                public RevertRevisionsCommand(RevertRevisionsRequest request)
-                {
-                    _request = request;
-                }
-
-                public override bool IsReadRequest => false;
-
-                public override HttpRequestMessage CreateRequest(JsonOperationContext ctx, ServerNode node, out string url)
-                {
-                    url = $"{node.Url}/databases/{node.Database}/revisions/revert";
-
-                    return new HttpRequestMessage
-                    {
-                        Method = HttpMethod.Post,
-                        Content = new BlittableJsonContent(async stream => await ctx.WriteAsync(stream, DocumentConventions.Default.Serialization.DefaultConverter.ToBlittable(_request, ctx)).ConfigureAwait(false))
-                    };
-                }
-
-                public override void SetResponse(JsonOperationContext context, BlittableJsonReaderObject response, bool fromCache)
-                {
-                    if (response == null)
-                        ThrowInvalidResponse();
-
-                    Result = DocumentConventions.Default.Serialization.DefaultConverter.FromBlittable<OperationIdResult>(response);
                 }
             }
         }
